+++ packages/core/src/entry.ts
@@ -2,6 +2,8 @@ import type Day from './types/Day';
 import type TimeSlot from './types/TimeSlot';
 import type { CustomField, CustomFieldCollected } from './types/InitData';
 import type ZaptimeConfig from './types/ZaptimeConfig';
+import type { ReservationResponse } from './types/ApiResponses';
+
 export { book, reserve, confirm, cancel, fetchRemoteConfiguration, reschedule } from './composables/useApi';
 export { default as useCalendar } from './composables/useCalendar';
 export { default as useConfig } from './composables/useConfig';
@@ -17,4 +19,4 @@ export { default as useReservationReschedule } from './composables/useReservatio
 
 export { default as mergeObjects } from './utils/mergeObjects';
 
-export type { Day, TimeSlot, ZaptimeConfig, CustomField, CustomFieldCollected };
+export type { Day, TimeSlot, ZaptimeConfig, CustomField, CustomFieldCollected, ReservationResponse };